refactor: improve error handling and centralize adapter messages
